Changed init_pollfd_record.

+++ accept.h
@@ -2,7 +2,7 @@
 #ifndef ACCEPT_HEADER_H_
 #define	ACCEPT_HEADER_H_
 
-int get_listening_socket(int server_port);
+int get_listening_socket(unsigned short server_port);
 void * run_accept_thread(void * arg);
 pthread_t create_accept_thread(int server_port);
